requirements.txt
asgiref==3.8.1
Django==5.0.6
django-model-utils==4.0.0
html5lib==1.1
pillow==10.4.0
psycopg2-binary==2.9.9
pypdf2==3.0.1
pytz==2022.1
reportlab==4.2.2
six
sqlparse==0.5.0
Unipath==1.1
webencodings==0.5.1
xhtml2pdf==0.2.16
django-extensions